chore: move myprofile page files to myaccount dir
M2-546
