+++ packages/theme/modules/customer/pages/MyAccount/MyProfile/ProfileUpdateForm.vue
@@ -113,7 +113,7 @@ import { defineComponent, ref } from '@nuxtjs/composition-api';
 import userGetters from '~/modules/customer/getters/userGetters';
 import { useUser } from '~/modules/customer/composables/useUser';
 import { useUiNotification } from '~/composables';
-import type { SubmitEventPayload, ProfileUpdateFormFields } from '~/modules/customer/pages/MyProfile/types';
+import type { SubmitEventPayload, ProfileUpdateFormFields } from '~/modules/customer/pages/MyAccount/MyProfile/types';
 
 extend('email', {
   ...email,
